# fix: prevent CCounter from halting due to very large BPM