--- a/OpenTaiko/src/Stages/07.Game/Taiko/CActImplMtaiko.cs
+++ b/OpenTaiko/src/Stages/07.Game/Taiko/CActImplMtaiko.cs
@@ -48,11 +48,11 @@ internal class CActImplMtaiko : CActivity {
 
 	public override int Draw() {
 		if (base.IsFirstDraw) {
-			this.nフラッシュ制御タイマ = (long)(SoundManager.PlayTimer.NowTime * OpenTaiko.ConfigIni.SongPlaybackSpeed);
+			this.nフラッシュ制御タイマ = (long)(SoundManager.PlayTimer.NowTimeMs * OpenTaiko.ConfigIni.SongPlaybackSpeed);
 			base.IsFirstDraw = false;
 		}
 
-		long num = (long)(SoundManager.PlayTimer.NowTime * OpenTaiko.ConfigIni.SongPlaybackSpeed);
+		long num = (long)(SoundManager.PlayTimer.NowTimeMs * OpenTaiko.ConfigIni.SongPlaybackSpeed);
 		if (num < this.nフラッシュ制御タイマ) {
 			this.nフラッシュ制御タイマ = num;
 		}