… OpenTaiko/src/Stages/07.Game/CStage演奏画面共通.cs
… OpenTaiko/src/Stages/07.Game/Taiko/CAct演奏Drumsゲームモード.cs
… OpenTaiko/src/Stages/07.Game/Taiko/CStage演奏ドラム画面.cs