This should not affect the game performance in average gameplay.

## Before

* 60fps-ish (debug build) in 愛歌「ネグラドルナ」(kirura969), halts at end.
* 480fps-ish in average gameplay (debug build)

## After

* 60fps-ish (debug build) in 愛歌「ネグラドルナ」(kirura969), no halts at end.

https://github.com/user-attachments/assets/2bc0057e-ee61-4cf3-b374-8098dead215b

* 480fps-ish in average gameplay (debug build)

IMO the difference of fps is not very significant.